+++ conf/example_apps/battery.py
@@ -1,5 +1,6 @@
 import appdaemon.appapi as appapi
 import datetime
+import globals
 
 #
 # App to send email report for devices running low on battery
@@ -8,7 +9,6 @@ import datetime
 #
 # threshold = value below which battery levels are reported and email is sent
 # always_send = set to 1 to override threshold and force send
-# input_select = Name of input_select to monitor followed by comma separated list of values for which on action should be performed
 #
 # None
 #
@@ -22,7 +22,7 @@ class Battery(appapi.AppDaemon):
   def initialize(self):
     #self.check_batteries({"force": 1})
     time = datetime.time(6, 0, 0)
-    self.run_daily(self.check_batteries, time) 
+    self.run_daily(self.check_batteries, time)
     
   def check_batteries(self, kwargs):
     devices = self.get_state()
@@ -30,14 +30,15 @@ class Battery(appapi.AppDaemon):
     low = []
     for device in devices:
       battery = None
-      if "battery" in devices[device]["attributes"]:
-        battery = devices[device]["attributes"]["battery"]
-      if "battery_level" in devices[device]["attributes"]:
-        battery = devices[device]["attributes"]["battery_level"]
-      if battery != None:
-        if battery < int(self.args["threshold"]):
-          low.append(device)
-        values[device] = battery
+      if "attributes" in devices[device]:
+          if "battery" in devices[device]["attributes"]:
+            battery = devices[device]["attributes"]["battery"]
+          if "battery_level" in devices[device]["attributes"]:
+            battery = devices[device]["attributes"]["battery_level"]
+          if battery != None:
+            if battery < int(self.args["threshold"]):
+              low.append(device)
+            values[device] = battery
     message = "Bettery Level Report\n\n"
     if low:
       message += "The following devices are low: (< {}) ".format(self.args["threshold"])
@@ -50,5 +51,5 @@ class Battery(appapi.AppDaemon):
       message += "{}: {}\n".format(device, values[device])
       
     if low or ("always_send" in self.args and self.args["always_send"] == "1") or ("force" in kwargs and kwargs["force"] == 1):
-      self.notify(message, title="Home Assistant Battery Report", name = "smtp")
+      self.notify(message, title="Home Assistant Battery Report", name = "andrew_mail")