+++ conf/example_apps/grandfather.py
@@ -4,6 +4,8 @@ import threading
 import time
 import os.path
 import datetime
+import globals
+
 #
 # Grandfather clock APP inspired by @areeshmu:
 #