+++ conf/example_apps/outside_lights.py
@@ -1,4 +1,5 @@
 import appdaemon.appapi as appapi
+import globals
 
 #
 # App to turn lights on and off at sunrise and sunset
@@ -20,8 +21,17 @@ class OutsideLights(appapi.AppDaemon):
     
   def sunrise_cb(self, kwargs):
     self.log("OutsideLights: Sunrise Triggered")
+    self.cancel_timers()
     self.turn_on(self.args["off_scene"])
 
   def sunset_cb(self, kwargs):
     self.log("OutsideLights: Sunset Triggered")
+    self.cancel_timers()
     self.turn_on(self.args["on_scene"])
+
+  def cancel_timers(self):
+    if "timers" in self.args:
+      apps = self.args["timers"].split(",")
+      for app in apps:
+        App = self.get_app(app)
+        App.cancel()