+++ conf/example_apps/smart_heat.py
@@ -1,5 +1,6 @@
 import appdaemon.appapi as appapi
 import datetime
+import globals
 
 #
 # App to manage heating: